requirements.txt
asgiref==3.3.4
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.4.0
django-filter==2.3.0
django-heroku==0.3.1
django-logging-middleware==0.1.2
django-oauth-toolkit==1.5.0
django-redis==4.12.1
djangorestframework==3.11.0
drf-yasg==1.17.1
flake8==3.8.4
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==0.8
loguru==0.5.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
psycopg2==2.8.6
psycopg2-binary==2.8.5
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
